+++ _examples/response-writer/sse-third-party/main.go
@@ -4,11 +4,11 @@ import (
 	"time"
 
 	"github.com/kataras/iris/v12"
-	"github.com/r3labs/sse"
+	"github.com/r3labs/sse/v2"
 )
 
 // First of all install the sse third-party package (you can use other if you don't like this approach or go ahead to the "sse" example)
-// $ go get -u github.com/r3labs/sse
+// $ go get github.com/r3labs/sse/v2@v2.7.4
 func main() {
 	app := iris.New()
 	s := sse.New()
@@ -22,7 +22,7 @@ func main() {
 	*/
 	s.CreateStream("messages")
 
-	app.Any("/events", iris.FromStd(s.HTTPHandler))
+	app.Any("/events", iris.FromStd(s))
 
 	go func() {
 		// You design when to send messages to the client,